% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.weightfunct.R
\name{print}
\alias{print}
\title{Print Model Results}
\usage{
print(x, ...)
}
\arguments{
\item{x}{an object of class weightfunct}

\item{...}{other arguments}
}
\description{
This function allows you to print the model results.
}
\examples{
\dontrun{
print(weightfunct(d,v))
}
}
\keyword{weightr}
