% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.weightfunct.R
\name{print.weightfunct}
\alias{print.weightfunct}
\title{Print Model Results}
\usage{
\method{print}{weightfunct}(x)
}
\description{
This function allows you to print the model results.
}
\examples{
\dontrun{
print.weightfunct(weightfunct(d,v))
}
}
\keyword{weightr}
