% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_weightr.R
\name{shiny_weightr}
\alias{shiny_weightr}
\title{Start weightr in Shiny}
\usage{
shiny_weightr()
}
\description{
This function allows you to launch the Shiny application locally.
}
\examples{
\dontrun{
library(shiny)
library(shinyBS)
shiny_weightr()
}
}
\keyword{weightr}

